backend/requirements.txt
#web framework libraries
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
#database libraries
sqlalchemy==2.0.23
aiosqlite==0.19.0
#security libraries
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-dotenv==1.0.0
#image processing libraries
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
#face recognition
face-recognition==1.3.0
dlib==19.24.2
#ocr and text reading
pytesseract==0.3.10
pdf2image==1.16.3
#liveliness detection
mediapipe==0.10.7
imutils==0.5.4
#ml for sorting and classification
scikit-learn==1.3.2
#utilities
pydantic==2.5.0
aiofiles==23.2.1
python-magic==0.4.27
requests==2.31.0
python-dateutil==2.8.2

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
